#18 fix: 메인 액티비티의 프래그먼트들 경로 변경
