+++ app/src/main/java/com/c7z/mappilogue_aos/presentation/ui/main/meet/MeetFragment.kt
@@ -1,4 +1,4 @@
-package com.c7z.mappilogue_aos.presentation.ui.mypage
+package com.c7z.mappilogue_aos.presentation.ui.main.meet
 
 import androidx.lifecycle.ViewModelProvider
 import android.os.Bundle
@@ -7,26 +7,26 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.c7z.mappilogue_aos.R
-import com.c7z.mappilogue_aos.presentation.ui.mypage.viewmodel.MypageViewModel
+import com.c7z.mappilogue_aos.presentation.ui.main.meet.viewmodel.MeetViewModel
 
-class MypageFragment : Fragment() {
+class MeetFragment : Fragment() {
 
     companion object {
-        fun newInstance() = MypageFragment()
+        fun newInstance() = MeetFragment()
     }
 
-    private lateinit var viewModel: MypageViewModel
+    private lateinit var viewModel: MeetViewModel
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        return inflater.inflate(R.layout.fragment_mypage, container, false)
+        return inflater.inflate(R.layout.fragment_meet, container, false)
     }
 
     override fun onActivityCreated(savedInstanceState: Bundle?) {
         super.onActivityCreated(savedInstanceState)
-        viewModel = ViewModelProvider(this).get(MypageViewModel::class.java)
+        viewModel = ViewModelProvider(this).get(MeetViewModel::class.java)
         // TODO: Use the ViewModel
     }
