+++ app/src/main/java/com/c7z/mappilogue_aos/presentation/ui/main/meet/viewmodel/MeetViewModel.kt
@@ -1,4 +1,4 @@
-package com.c7z.mappilogue_aos.presentation.ui.meet.viewmodel
+package com.c7z.mappilogue_aos.presentation.ui.main.meet.viewmodel
 
 import androidx.lifecycle.ViewModel
