+++ app/src/main/java/com/c7z/mappilogue_aos/presentation/ui/main/mypage/MypageFragment.kt
@@ -1,4 +1,4 @@
-package com.c7z.mappilogue_aos.presentation.ui.meet
+package com.c7z.mappilogue_aos.presentation.ui.main.mypage
 
 import androidx.lifecycle.ViewModelProvider
 import android.os.Bundle
@@ -7,26 +7,26 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.c7z.mappilogue_aos.R
-import com.c7z.mappilogue_aos.presentation.ui.meet.viewmodel.MeetViewModel
+import com.c7z.mappilogue_aos.presentation.ui.main.mypage.viewmodel.MypageViewModel
 
-class MeetFragment : Fragment() {
+class MypageFragment : Fragment() {
 
     companion object {
-        fun newInstance() = MeetFragment()
+        fun newInstance() = MypageFragment()
     }
 
-    private lateinit var viewModel: MeetViewModel
+    private lateinit var viewModel: MypageViewModel
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        return inflater.inflate(R.layout.fragment_meet, container, false)
+        return inflater.inflate(R.layout.fragment_mypage, container, false)
     }
 
     override fun onActivityCreated(savedInstanceState: Bundle?) {
         super.onActivityCreated(savedInstanceState)
-        viewModel = ViewModelProvider(this).get(MeetViewModel::class.java)
+        viewModel = ViewModelProvider(this).get(MypageViewModel::class.java)
         // TODO: Use the ViewModel
     }
