+++ app/src/main/res/layout/fragment_memorize.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/gray_1"
-    tools:context=".presentation.ui.memorize.MemorizeFragment">
+    tools:context=".presentation.ui.main.memorize.MemorizeFragment">
 
     <TextView
         android:layout_width="wrap_content"
