+++ app/src/main/res/layout/fragment_mypage.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/gray_1"
-    tools:context=".presentation.ui.mypage.MypageFragment">
+    tools:context=".presentation.ui.main.mypage.MypageFragment">
 
     <TextView
         android:layout_width="wrap_content"
